+++ packages/admin/admin/src/form/file/FileSelect.tsx
@@ -9,12 +9,14 @@ import { Alert } from "../../alert/Alert";
 import { createComponentSlot } from "../../helpers/createComponentSlot";
 import { ThemedComponentBaseProps } from "../../helpers/ThemedComponentBaseProps";
 import { FileDropzone } from "./FileDropzone";
-import { FileSelectItem, ValidFileSelectItem } from "./fileSelectItemTypes";
+import { ErrorFileSelectItem, FileSelectItem, ValidFileSelectItem } from "./fileSelectItemTypes";
 import { FileSelectListItem } from "./FileSelectListItem";
 import { getFilesInfoText } from "./getFilesInfoText";
 
 export type FileSelectClassKey =
     | "root"
+    | "listLayout"
+    | "gridLayout"
     | "maxFilesReachedInfo"
     | "dropzone"
     | "fileList"
@@ -22,6 +24,8 @@ export type FileSelectClassKey =
     | "error"
     | "errorMessage"
     | "filesInfoText";
+
+type Layout = "list" | "grid";
 
 type ThemeProps = ThemedComponentBaseProps<{
     root: "div";
@@ -34,11 +38,15 @@ type ThemeProps = ThemedComponentBaseProps<{
     filesInfoText: typeof FormHelperText;
 }>;
 
+type OwnerState = {
+    layout: Layout;
+};
+
 export type FileSelectProps<AdditionalValidFileValues = Record<string, unknown>> = {
     files: FileSelectItem<AdditionalValidFileValues>[];
     onDrop?: DropzoneOptions["onDrop"];
-    onRemove?: (file: FileSelectItem<AdditionalValidFileValues>) => void;
-    onDownload?: (file: FileSelectItem<AdditionalValidFileValues>) => void;
+    onRemove?: (file: ValidFileSelectItem<AdditionalValidFileValues> | ErrorFileSelectItem) => void;
+    onDownload?: (file: ValidFileSelectItem<AdditionalValidFileValues>) => void;
     getDownloadUrl?: (file: ValidFileSelectItem<AdditionalValidFileValues>) => string;
     disabled?: boolean;
     readOnly?: boolean;
@@ -47,6 +55,7 @@ export type FileSelectProps<AdditionalValidFileValues = Record<string, unknown>>
     maxFiles?: number;
     multiple?: boolean;
     error?: React.ReactNode;
+    layout?: Layout;
     iconMapping?: {
         error?: React.ReactNode;
     };
@@ -68,6 +77,7 @@ export const FileSelect = <AdditionalValidFileValues = Record<string, unknown>,>
         getDownloadUrl,
         files,
         error,
+        layout = "list",
         ...restProps
     } = useThemeProps({
         props: inProps,
@@ -86,8 +96,12 @@ export const FileSelect = <AdditionalValidFileValues = Record<string, unknown>,>
     const filesInfoText = getFilesInfoText(maxFiles, maxFileSize);
     const showFileList = files.length > 0 || readOnly;
 
+    const ownerState: OwnerState = {
+        layout,
+    };
+
     return (
-        <Root {...slotProps?.root} {...restProps}>
+        <Root ownerState={ownerState} {...slotProps?.root} {...restProps}>
             {!readOnly && (
                 <>
                     {maxAmountOfFilesSelected ? (
@@ -116,7 +130,7 @@ export const FileSelect = <AdditionalValidFileValues = Record<string, unknown>,>
                 </>
             )}
             {showFileList && (
-                <FileList {...slotProps?.fileList}>
+                <FileList ownerState={ownerState} {...slotProps?.fileList}>
                     {files.length > 0 ? (
                         <>
                             {files.map((file, index) => {
@@ -134,7 +148,8 @@ export const FileSelect = <AdditionalValidFileValues = Record<string, unknown>,>
                                                 : undefined
                                         }
                                         downloadUrl={isValidFile && getDownloadUrl ? getDownloadUrl(file) : undefined}
-                                        onClickDelete={readOnly || !onRemove ? undefined : () => onRemove(file)}
+                                        onClickDelete={readOnly || !onRemove || "loading" in file ? undefined : () => onRemove(file)}
+                                        filePreview={layout === "grid"}
                                         {...slotProps?.fileListItem}
                                     />
                                 );
@@ -166,9 +181,10 @@ export const FileSelect = <AdditionalValidFileValues = Record<string, unknown>,>
     );
 };
 
-const Root = createComponentSlot("div")<FileSelectClassKey>({
+const Root = createComponentSlot("div")<FileSelectClassKey, OwnerState>({
     componentName: "FileSelect",
     slotName: "root",
+    classesResolver: ({ layout }) => [layout === "list" && "listLayout", layout === "grid" && "gridLayout"],
 })(css`
     display: flex;
     flex-direction: column;
@@ -181,12 +197,29 @@ const Dropzone = createComponentSlot(FileDropzone)<FileSelectClassKey>({
     slotName: "dropzone",
 })();
 
-const FileList = createComponentSlot("div")<FileSelectClassKey>({
+const FileList = createComponentSlot("div")<FileSelectClassKey, OwnerState>({
     componentName: "FileSelect",
     slotName: "fileList",
-})(css`
-    width: 100%;
-`);
+})(
+    ({ theme, ownerState }) => css`
+        width: 100%;
+
+        ${ownerState.layout === "grid" &&
+        css`
+            display: grid;
+            gap: ${theme.spacing(2)};
+            grid-template-columns: repeat(auto-fill, minmax(150px, 1fr));
+
+            ${theme.breakpoints.up("md")} {
+                grid-template-columns: repeat(auto-fill, minmax(200px, 1fr));
+
+                .CometAdminFormFieldContainer-horizontal & {
+                    grid-template-columns: repeat(auto-fill, minmax(150px, 1fr));
+                }
+            }
+        `}
+    `,
+);
 
 const FileListItem = createComponentSlot(FileSelectListItem)<FileSelectClassKey>({
     componentName: "FileSelect",