+++ packages/admin/admin-theme/CHANGELOG.md
@@ -218,6 +218,18 @@
 
 -   @comet/admin-icons@7.0.0-beta.0
 
+## 6.17.1
+
+### Patch Changes
+
+-   @comet/admin-icons@6.17.1
+
+## 6.17.0
+
+### Patch Changes
+
+-   @comet/admin-icons@6.17.0
+
 ## 6.16.0
 
 ### Patch Changes